ESP32 Phase/csi-tracking-system/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
ultralytics>=8.3.0
filterpy>=1.4.5
matplotlib>=3.7.0
pyyaml>=6.0
pyserial>=3.5
pandas>=2.0.0
